Added documentation for the integrators
