+++ src/us/stomberg/solarsystemsim/Main.java
@@ -2,8 +2,6 @@ package us.stomberg.solarsystemsim;
 
 import us.stomberg.solarsystemsim.graphics.DrawSpace;
 import us.stomberg.solarsystemsim.physics.Physics;
-import us.stomberg.solarsystemsim.physics.ExplicitEulerIntegrator;
-import us.stomberg.solarsystemsim.physics.SymplecticIntegrator;
 import us.stomberg.solarsystemsim.physics.VerletIntegrator;
 
 /**
